+++ acceptance/help_test.go
@@ -97,7 +97,7 @@ Global Flags:
   --trace, -tr, OM_TRACE                                 bool    prints HTTP requests and response payloads
   --username, -u, OM_USERNAME                            string  admin username for the Ops Manager VM (not required for unauthenticated commands)
   --version, -v                                          bool    prints the om release version (default: false)
-  OM_VARS_ENV                                            string  **EXPERIMENTAL** load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)
+  OM_VARS_ENV                                            string  load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)
 `
 
 const CONFIGURE_AUTHENTICATION_USAGE = `
@@ -116,6 +116,6 @@ Flags:
   --precreated-client-secret                              string             create a UAA client on the Ops Manager vm. The client_secret will be the value provided to this option
   --username, -u, OM_USERNAME                             string (required)  admin username
   --var, -v                                               string (variadic)  load variable from the command line. Format: VAR=VAL
-  --vars-env, OM_VARS_ENV                                 string (variadic)  **EXPERIMENTAL** load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)
+  --vars-env, OM_VARS_ENV                                 string (variadic)  load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)
   --vars-file, -l                                         string (variadic)  load variables from a YAML file
 ` + GLOBAL_USAGE_FLAGS