+++ commands/bosh_diff.go
@@ -173,8 +173,8 @@ func (c BoshDiff) colorizeDiff(diff string) string {
 
 func (c BoshDiff) Usage() jhanda.Usage {
 	return jhanda.Usage{
-		Description:      "**EXPERIMENTAL** This command displays the bosh manifest diff for the director and products (Note: secret values are replaced with double-paren variable names)",
-		ShortDescription: "**EXPERIMENTAL** displays BOSH manifest diff for the director and products",
+		Description:      "This command displays the bosh manifest diff for the director and products (Note: secret values are replaced with double-paren variable names)",
+		ShortDescription: "displays BOSH manifest diff for the director and products",
 		Flags:            c.Options,
 	}
 }