+++ commands/configure_saml_authentication.go
@@ -24,7 +24,7 @@ type ConfigureSAMLAuthentication struct {
 		RBACGroupsAttribute       string   `long:"saml-rbac-groups-attribute"       required:"true" description:"If SAML is specified, please provide the groups attribute for your SAML"`
 		SkipCreateBoshAdminClient bool     `long:"skip-create-bosh-admin-client"                    description:"create a UAA client on the Bosh Director, whose credentials can be passed to the BOSH CLI to execute BOSH commands. Default is false."`
 		PrecreatedClientSecret    string   `long:"precreated-client-secret"                         description:"create a UAA client on the Ops Manager vm, whose secret will be the value provided to this option"`
-		VarsEnv                   []string `long:"vars-env" env:"OM_VARS_ENV" experimental:"true"   description:"load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)"`
+		VarsEnv                   []string `long:"vars-env" env:"OM_VARS_ENV"                       description:"load vars from environment variables by specifying a prefix (e.g.: 'MY' to load MY_var=value)"`
 		VarsFile                  []string `long:"vars-file"             short:"l"                  description:"load variables from a YAML file"`
 		Vars                      []string `long:"var"                   short:"v"                  description:"load variable from the command line. Format: VAR=VAL"`
 	}